% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array2list}
\alias{array2list}
\title{\code{array2list}: converts three dimensional arrays to a list of matrices}
\usage{
array2list(x)
}
\arguments{
\item{x}{An array to convert to a list of matrices}
}
\value{
A boolean. TRUE if the matrix is a square and symmetric matrix, FALSE if the matrix is not.
}
\description{
\code{array2list} takes a three dimensional array and converts it to a list of length \code{dim(x)[[3]]},
where each matrix in the list has \code{dim(x)[[1]]} rows and  \code{dim(x)[[1]]} columns.
}
\note{
This function is copied directly from https://stackoverflow.com/questions/20198751/three-dimensional-array-to-list
}
