% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_alphas.R
\name{compute_alphas}
\alias{compute_alphas}
\title{compute_alphas: compute alpha weights per matrix.}
\usage{
compute_alphas(centered_normed_matrices, alpha_from_RV = TRUE)
}
\arguments{
\item{centered_normed_matrices}{list. These must be the normalized cross-product matrices, i.e., they have already been double centered then normalzied.}

\item{alpha_from_RV}{boolean. Default is TRUE. If TRUE, this scales the vectorized \code{centered_normed_matrices} and provides weights based on the Rv-coefficient. Else (FALSE) the vectorized matrices are only centered.}
}
\value{
vector. These are the weights per matrix to make the compromise.
}
\description{
compute_alphas: compute alpha weights per matrix.
}
