% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_barycentric_partial_component_scores.R
\name{compute_barycentric_partial_component_scores}
\alias{compute_barycentric_partial_component_scores}
\title{compute_barycentric_partial_component_scores: compute re-weighted versions of \code{partial_component_scores} which are barycentric (i.e., equal the compromise scores)}
\usage{
compute_barycentric_partial_component_scores(partial_component_scores,
  alpha_weights)
}
\arguments{
\item{partial_component_scores}{list. A list of the partial component scores as computed from \code{compute_partial_component_scores}}

\item{alpha_weights}{vector. A vector that contains the alpha weights as used in the computation of the compromise.}
}
\value{
list. A list of the barycentric version of the partial component scores.
}
\description{
compute_barycentric_partial_component_scores: compute re-weighted versions of \code{partial_component_scores} which are barycentric (i.e., equal the compromise scores)
}
\details{
The name "barycentric partial component scores" means that the partial component scores (\code{partial_component_scores}) are re-weighted by
the number of tables (\code{length(alpha_weights)}) and the alpha weights (\code{alpha_weights}). The mean of the barycentric partial component scores
is equal to the compromise component scores.
}
\examples{
TODO
}
