% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_partial_component_scores.R
\name{compute_partial_component_scores}
\alias{compute_partial_component_scores}
\title{compute_partial_component_scores: compute the partial component scores (per matrix) for the compromise space}
\usage{
compute_partial_component_scores(centered_normed_matrices,
  compromise_decomposition_results)
}
\arguments{
\item{centered_normed_matrices}{list. These must be the normalized cross-product matrices, i.e., they have already been double centered then normalzied.}

\item{compromise_decomposition_results}{list. These must be the results from the decomposition of the compromise matrix (e.g., \code{$vectors} and \code{$values}).}
}
\value{
list. Each element in the list is a matrix of partial component scores.
}
\description{
compute_partial_component_scores: compute the partial component scores (per matrix) for the compromise space
}
