% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_relative_distance.R
\name{compute_relative_distance}
\alias{compute_relative_distance}
\title{Compute individual component scores relative to compromise component score.}
\usage{
compute_relative_distance(covstatis_results)
}
\arguments{
\item{covstatis_results}{TODO}
}
\value{
TODO
}
\description{
TODO
}
