% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covstatis.R
\docType{package}
\name{CovSTATIS-package}
\alias{CovSTATIS}
\alias{CovSTATIS-package}
\title{CovSTATIS}
\description{
Perhaps the statement here is not malformed anymore.
}
\details{
not malformed anymored.
}
\examples{
\dontrun{
## hello
}

}
\references{
Abdi, H., Williams, L.J., Valentin, D., & Bennani-Dosse, M. (2012). STATIS and DISTATIS: Optimum multi-table principal component analysis and three way metric multidimensional scaling. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 4, 124-167.
}
\seealso{
\code{\link{GSVD}}
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

Authors:
\itemize{
  \item Jenny Rieck
}

Other contributors:
\itemize{
  \item Carina Fan [contributor]
}

}
\keyword{component}
\keyword{decomposition}
\keyword{diagonalization}
\keyword{eigen}
\keyword{generalized}
\keyword{matrix}
\keyword{multivariate}
\keyword{orthogonal}
\keyword{svd}
\keyword{variance}
