% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_center_matrices.R
\name{double_center_a_matrix}
\alias{double_center_a_matrix}
\title{double_center_a_matrix: double center a matrix}
\usage{
double_center_a_matrix(sspsd_matrix)
}
\arguments{
\item{sspsd_matrix}{matrix. A square, symmetric, and (presumed) positive semi-definite---i.e., covariance or correlation---matrix to be centered}
}
\value{
matrix. A double centered covariance or correlation matrix
}
\description{
double_center_a_matrix: double center a matrix
}
