% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_center_matrices.R
\name{double_center_matrices}
\alias{double_center_matrices}
\title{double_center_matrices: double center a list of matrices}
\usage{
double_center_matrices(cov_matrices)
}
\arguments{
\item{cov_matrices}{list. Covariance or correlation matrices to be double centered}
}
\value{
list. Double centered covariance or correlation matrices
}
\description{
double_center_matrices: double center a list of matrices
}
