% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eigen_tolerance}
\alias{eigen_tolerance}
\title{Eigendecomposition tolerance corrections}
\usage{
eigen_tolerance(eigen_results, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{tol}{Tolerance precision to eliminate all abs(x) values below \code{tol}. Default is \code{1e-12}.}

\item{eigen_tolerance}{results from \code{eigen()}}
}
\value{
\code{eigen_tolerance} corrected for tolerance (if necessary)
}
\description{
takes the results of an eigen decomposition and drops vectors and values below
tolerance threshold
}
\seealso{
\code{\link{eigen}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
