% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_compromise_matrix.R
\name{make_compromise_matrix}
\alias{make_compromise_matrix}
\title{make_compromise_matrix: make the compromise (weighted average) matrix}
\usage{
make_compromise_matrix(centered_normed_matrices, alpha_weights)
}
\arguments{
\item{centered_normed_matrices}{list. A list of covariance or correlation matrices that have been double centered and normalized}

\item{alpha_weights}{vector. A vector of weights to use for the weighted average to produce the compromise matrix}
}
\value{
matrix. The weighted average of all matrices in \code{centered_normed_matrices} with the weights of \code{alpha_weights}
}
\description{
make_compromise_matrix: make the compromise (weighted average) matrix
}
