% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrices.R
\name{normalize_a_matrix}
\alias{normalize_a_matrix}
\title{normalize_a_matrix: normalize a double centered matrix}
\usage{
normalize_a_matrix(centered_matrix, matrix_norm_type = "MFA")
}
\arguments{
\item{centered_matrix}{matrix. A double centered matrix to be normalized}

\item{matrix_norm_type}{character string. Normalization type for \code{centered_matrix}. Options are "MFA", "SS1", and "none"}
}
\value{
list. A list of normalized matrices
}
\description{
normalize_a_matrix: normalize a double centered matrix
}
