% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrices.R
\name{normalize_matrices}
\alias{normalize_matrices}
\title{normalize_matrices: normalize a list of matrices}
\usage{
normalize_matrices(centered_matrices, matrix_norm_type = "MFA")
}
\arguments{
\item{centered_matrices}{list. A list of double centered matrices to be normalized.}

\item{matrix_norm_type}{character string. Normalization type for each \code{centered_matrices}. Options are "MFA", "SS1", and "none"}
}
\value{
list. A list of normalized matrices
}
\description{
normalize_matrices: normalize a list of matrices
}
