% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covstatis.R
\name{plot_covstatis}
\alias{plot_covstatis}
\title{TODO: TITLE}
\usage{
plot_covstatis(fi, fis, fi.cols = NULL, fis.cols = NULL, axes = c(1,
  2), display.fi.names = T, display.fis.names = F)
}
\arguments{
\item{fi}{TODO}

\item{fis}{TODO}

\item{fi.cols}{TODO. Default is NULL.}

\item{fis.cols}{TODO. Default is NULL.}

\item{axes}{TODO. Default is 1 and 2.}

\item{display.fi.names}{TODO. Default is TRUE.}

\item{display.fis.names}{TODO.}

\item{fis.names}{TODO. Default is FALSE.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
## hello
}
}
