% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_low_rank.R
\name{rebuild_low_rank}
\alias{rebuild_low_rank}
\title{Rebuild low-rank matrices for compromise and each individual matrix
NEEDS CHECKING/COMPLETION/TESTING.}
\usage{
rebuild_low_rank(covstatis_results, components_to_rebuild)
}
\arguments{
\item{covstatis_results}{TODO}

\item{components_to_rebuild}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}
