% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distatis.R
\name{distatis}
\alias{distatis}
\title{TODO: TITLE}
\usage{
distatis(squared_dist_matrices, matrix_norm_type = "MFA",
  alpha_from_RV = TRUE, tolerance = sqrt(.Machine$double.eps),
  strictly_enforce_psd = FALSE)
}
\arguments{
\item{squared_dist_matrices}{TODO}

\item{matrix_norm_type}{TODO. Default is MFA.}

\item{alpha_from_RV}{TODO. Default is TRUE.}

\item{compact}{TODO. Default is TRUE.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}
